restructure files, assign member to their own file
