+++ ui_handling.py
@@ -1,0 +1,6 @@
+#Oksana
+
+
+def print_hangman(tries: int) -> None:
+    """Prints the hangman figure based on the number of tries left."""
+    raise NotImplementedError("This function is not implemented yet.")
