--- words
@@ -1,21 +1,0 @@
-word_list = [
-    "Laptop", 
-    "Pencil",
-    "Mountain",
-    "River",
-    "Book",
-    "Bridge",
-    "Carpet",
-    "Elephant",
-    "Galaxy",
-    "Keyboard",
-    "Ocean",
-    "Violin",
-    "Basketball",
-    "Clock",
-    "Garden",
-    "Sculpture",
-    "Helicopter",
-    "Volcano",
-    "Factory", 
-    "Train"]
